feat(ctb): Perform deployment using gnosis safe as system owner
temp: Add instrumentation to deploy.sol

fix(ctb): Use correct names for deployments

We deploy a SafeProxy contract as the system owner, it needs to be correctly named
in order to connect it to the source code.

The SafeProxyFactory and Safe (as the singleton) are also now saved.

fix(ctb): Skip HH deployments if deployTx not found

At the moment this only applies to the Safe contract, which is deployed in
a nested create2 call. We don't expect we'll need this artifact as it's not a
core part of the OP Stack.